requirements.txt
python-telegram-bot
bs4
lxml
requests
